Move direct properties configuration code into public functions in scripting host jar
